Add files via upload
For the app.py and the index.html in the app folder - Please be reminded to change the codes inside ( Add into it so to integrate your page into the app.py). Basically ur app.py stuff will be in the app.py and the index.html just change your list of features name? I think the format works HAHA.